Do not use e3 api to reset perspective
It doesn't know about e4 parts and thus closes the editors.
Change the Main perspective views to be fully defined in e4 xmi rather
than partially through code.
Fixes https://github.com/eclipse-chemclipse/chemclipse/issues/901 with
the limitation that if a view has been moved it stays in the new
position. If the view has been closed it's restored in its proper place.